Re-organize image tool files to avoid Qt dependencies outside of qt/ sub-directories
