+++ glue/viewers/image/pixel_selection_subset_state.py
@@ -1,0 +1,10 @@
+from __future__ import absolute_import, division, print_function
+
+from glue.core.subset import SliceSubsetState
+
+__all__ = ['PixelSubsetState']
+
+
+class PixelSubsetState(SliceSubsetState):
+    def copy(self):
+        return PixelSubsetState(self.reference_data, self.slices)
