+++ glue/viewers/image/qt/contrast_mouse_mode.py
